requirements.txt
# === Digest / RSS ===
feedparser>=6.0,<7
requests>=2.31,<3
pyyaml>=6.0,<7
beautifulsoup4>=4.12,<5
lxml>=5.0,<6
jinja2>=3.1,<4
markdown>=3.5,<4

# === X Post (Gemini + Tweepy) ===
google-genai>=1.0.0
tweepy>=4.14.0
